backend/requirements.txt
asgiref==3.11.1
dj-database-url==3.1.2
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.1.0
packaging==26.0
psycopg==3.3.3
psycopg-binary==3.3.3
PyJWT==2.11.0
sqlparse==0.5.5
typing_extensions==4.15.0
whitenoise==6.12.0
